lep8 escape code with avx512

+++ include/pdx/distance_computers/avx512_computers.hpp
@@ -775,8 +775,8 @@ public:
                     }
                     else {
                         //std::cout << "Patching nibble 0 " << "\n";
-                        float diff_high = exceptions_query[dimension_idx] - (float)(exceptions_data[exc_start_0 + exc_offset_0]);
-                        distances_p[vector_idx] += diff_high * diff_high * scaling_factors_exceptions[dimension_idx];
+                        float diff_high = exceptions_query[dimension_idx] - exceptions_data[exc_start_0 + exc_offset_0];
+                        distances_p[vector_idx] += (diff_high * diff_high * scaling_factors_exceptions[dimension_idx]);
                         exc_offset_0 += 1;
                     }
                     if (nibble_1 != EXC_ESCAPE_CODE_SCALAR) {
@@ -784,8 +784,8 @@ public:
                         distances_p[vector_idx] += diff_low * diff_low * scale_1;
                     }
                     else {
-                        float diff_low = exceptions_query[dimension_idx + 1] - (float)(exceptions_data[exc_start_1 + exc_offset_1]);
-                        distances_p[vector_idx] += diff_low * diff_low * scaling_factors_exceptions[dimension_idx + 1];
+                        float diff_low = exceptions_query[dimension_idx + 1] - exceptions_data[exc_start_1 + exc_offset_1];
+                        distances_p[vector_idx] += (diff_low * diff_low * scaling_factors_exceptions[dimension_idx + 1]);
                         exc_offset_1 += 1;
                     }
                 }
